+++ tests/e2e/support/index.js
@@ -1,0 +1,10 @@
+import './signOut';
+import './signInAdminUser';
+import './signedInAsAdminUser';
+import './deleteTestDepartment';
+import './createTestDepartment';
+import './createProduct';
+import './deleteProduct';
+import './createPeriod';
+import './createObjective';
+import './createKeyres';